requirements.txt
altgraph==0.17.4
blinker==1.8.2
bottle==0.13.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.15
dash-html-components==2.0.0
dash-table==5.0.0
dash_ag_grid==31.3.0
dataclass-wizard==0.22.3
EditorConfig==0.12.4
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Caching==2.3.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
macholib==1.16.3
MarkupSafe==2.1.5
more-itertools==9.1.0
nest-asyncio==1.6.0
numpy==1.24.4
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
plotly==5.24.1
proxy_tools==0.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-Security==10.3.2
pyobjc-framework-WebKit==10.3.2
python-dateutil==2.9.0.post0
pytz==2024.2
pywebview==5.3.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
